refactor: change property type from Double to BigDecimal

+++ backend/src/main/java/br/com/financeai/dto/response/FinancialAnalysisResponse.java
@@ -2,11 +2,15 @@ package br.com.financeai.dto.response;
 
 import br.com.financeai.enums.FinancialProfile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record FinancialAnalysisResponse(
         FinancialProfile perfilFinanceiro,
-        Double probabilidade,
+
+        BigDecimal probabilidade,
+
         ExpenseSummaryResponse resumoGastos,
+
         List<String> recomendacoes
 ) {}
