+++ backend/src/main/java/br/com/financeai/integration/dto/MlResponse.java
@@ -3,12 +3,13 @@ package br.com.financeai.integration.dto;
 import br.com.financeai.dto.response.ExpenseSummaryResponse;
 import br.com.financeai.enums.FinancialProfile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record MlResponse(
         FinancialProfile perfilFinanceiro,
 
-        Double probabilidade,
+        BigDecimal probabilidade,
 
         ExpenseSummaryResponse resumoGastos,
